Move the location of the popular-reviews
