+++ packages/review/src/components/reviews-shorten.tsx
@@ -9,7 +9,7 @@ import { formatNumber } from '@titicaca/view-utilities'
 import { useReviewCount } from '../services'
 
 import { LatestReviews } from './latest-reviews'
-import { PopularReviews } from './popular-reviews'
+import { PopularReviews } from './shorten-list'
 import { WriteButton } from './write-button'
 import { FilterProvider, useReviewFilters } from './filter-context'
 import {
