+++ packages/review/src/components/shorten-list/index.ts
@@ -1,0 +1,1 @@
+export * from './popular-reviews'
